+++ test/hexapdf/document/test_signatures.rb
@@ -59,15 +59,15 @@ describe HexaPDF::Document::Signatures do
     end
 
     it "can sign the data using PKCS7" do
-      data = "data"
+      data = StringIO.new("data")
       store = OpenSSL::X509::Store.new
       store.add_cert(CERTIFICATES.ca_certificate)
 
-      pkcs7 = OpenSSL::PKCS7.new(@handler.sign(data))
+      pkcs7 = OpenSSL::PKCS7.new(@handler.sign(data, [0, 4, 0, 0]))
       assert(pkcs7.detached?)
       assert_equal([CERTIFICATES.signer_certificate, CERTIFICATES.ca_certificate],
                    pkcs7.certificates)
-      assert(pkcs7.verify([], store, data, OpenSSL::PKCS7::DETACHED | OpenSSL::PKCS7::BINARY))
+      assert(pkcs7.verify([], store, data.string, OpenSSL::PKCS7::DETACHED | OpenSSL::PKCS7::BINARY))
     end
 
     describe "finalize_objects" do