+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * [HexaPDF::Document::Signatures::DefaultHandler] to allow setting the used
   signature method
+* **Breaking change**: [HexaPDF::Document::Signatures::DefaultHandler#sign]
+  needs to accept the IO object and the byte range instead of just the data
 
 
 ## 0.26.2 - 2022-10-22